+++ testdata/tree/with_child_directories/src/module/unreferenced_mod.rs
@@ -1,0 +1,5 @@
+//! This source file is not referenced by any `mod` statement, so it should not be visited.
+
+fn untested() -> i32 {
+    32
+}